refactor(service-sg-server): change import paths
https://github.com/akky-xxxx/project-my-homepage/issues/44
